requirements.txt
asgiref==3.4.1
Django==3.2.16
django-environ==0.9.0
Pillow==8.4.0

python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.2
whitenoise==5.3.0
psycopg2-binary==2.9.5
